Restructure subscription tests into modular directory layout
Move graphql_subscription_tests.rs into graphql_subscription/ directory:
- testing/harness.rs: shared SubscriptionTestCluster and SubscriptionStream
- checkpoint_subscription.rs: checkpoint field tests
