+++ crates/sui-indexer-alt-e2e-tests/tests/graphql_subscription/checkpoint_subscription.rs
@@ -1,0 +1,62 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+use sui_macros::sim_test;
+use test_cluster::TestClusterBuilder;
+
+use super::testing::*;
+
+#[sim_test]
+async fn test_subscription_sequential() {
+    let validator_cluster = TestClusterBuilder::new()
+        .with_num_validators(1)
+        .build()
+        .await;
+    let cluster = SubscriptionTestCluster::new(&validator_cluster).await;
+
+    let mut stream = cluster
+        .subscribe("subscription { checkpoints { sequenceNumber } }")
+        .await;
+    let items = stream.collect_items(3).await;
+
+    insta::assert_json_snapshot!("subscription_sequential", items);
+}
+
+#[sim_test]
+async fn test_subscription_fields() {
+    let validator_cluster = TestClusterBuilder::new()
+        .with_num_validators(1)
+        .build()
+        .await;
+    let cluster = SubscriptionTestCluster::new(&validator_cluster).await;
+
+    let mut stream = cluster
+        .subscribe(
+            r#"subscription {
+                checkpoints {
+                    sequenceNumber
+                    digest
+                    contentDigest
+                    timestamp
+                    networkTotalTransactions
+                    rollingGasSummary {
+                        computationCost
+                        storageCost
+                        storageRebate
+                        nonRefundableStorageFee
+                    }
+                    epoch {
+                        epochId
+                    }
+                    validatorSignatures {
+                        signature
+                        signersMap
+                    }
+                }
+            }"#,
+        )
+        .await;
+    let item = stream.next_item().await;
+
+    insta::assert_json_snapshot!("subscription_fields", item);
+}
