+++ crates/sui-indexer-alt-e2e-tests/tests/graphql_subscription/main.rs
@@ -1,0 +1,6 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+mod testing;
+
+mod checkpoint_subscription;
