+++ crates/sui-indexer-alt-e2e-tests/tests/graphql_subscription/testing/mod.rs
@@ -1,0 +1,6 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+mod harness;
+
+pub use harness::*;
